Start factoring out common parts of the FFI and JNI libraries
