+++ rust/bridge/ffi/Cargo.toml
@@ -17,6 +17,7 @@ crate-type = ["staticlib"]
 [dependencies]
 libsignal-protocol-rust = { path = "../../protocol" }
 aes-gcm-siv = { path = "../../aes-gcm-siv" }
+libsignal-bridge = { path = "../shared", features = ["ffi"] }
 async-trait = "0.1.41"
 libc = "0.2"
 futures = "0.3.7"
@@ -25,4 +26,4 @@ static_assertions = "1.1"
 log = "0.4.11"
 
 [build-dependencies]
-cbindgen = "0.14"
+cbindgen = "0.15"
