+++ rust/bridge/ffi/cbindgen.toml
@@ -39,6 +39,7 @@ sort_by = "None"
 [parse]
 parse_deps = true
 include = ["libsignal-protocol-rust", "aes-gcm-siv"]
+extra_bindings = ["libsignal-bridge"]
 
 [parse.expand]
-crates = ["libsignal-ffi"]
+crates = ["libsignal-ffi", "libsignal-bridge"]
