+++ rust/bridge/ffi/src/util.rs
@@ -6,25 +6,13 @@
 use futures::pin_mut;
 use futures::task::noop_waker_ref;
 use libc::{c_char, c_uchar, c_uint, c_ulonglong, size_t};
+use libsignal_bridge::*;
 use libsignal_protocol_rust::*;
 use std::ffi::{CStr, CString};
-use std::fmt;
 use std::future::Future;
 use std::task::{self, Poll};
 
 use aes_gcm_siv::Error as AesGcmSivError;
-
-#[derive(Debug)]
-pub enum SignalFfiError {
-    Signal(SignalProtocolError),
-    AesGcmSiv(AesGcmSivError),
-    InsufficientOutputSize(usize, usize),
-    NullPointer,
-    InvalidUtf8String,
-    UnexpectedPanic(std::boxed::Box<dyn std::any::Any + std::marker::Send>),
-    CallbackError(i32),
-    InvalidType,
-}
 
 #[derive(Debug)]
 #[repr(C)]
@@ -152,58 +140,6 @@ impl From<&SignalFfiError> for SignalErrorCode {
 
             _ => SignalErrorCode::UnknownError,
         }
-    }
-}
-
-impl fmt::Display for SignalFfiError {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match self {
-            SignalFfiError::Signal(s) => write!(f, "{}", s),
-            SignalFfiError::CallbackError(c) => {
-                write!(f, "callback invocation returned error code {}", c)
-            }
-            SignalFfiError::AesGcmSiv(c) => {
-                write!(f, "AES-GCM-SIV operation failed: {}", c)
-            }
-            SignalFfiError::NullPointer => write!(f, "null pointer"),
-            SignalFfiError::InvalidType => write!(f, "invalid type"),
-            SignalFfiError::InvalidUtf8String => write!(f, "invalid UTF8 string"),
-            SignalFfiError::InsufficientOutputSize(n, h) => {
-                write!(f, "needed {} elements only {} provided", n, h)
-            }
-
-            SignalFfiError::UnexpectedPanic(e) => match e.downcast_ref::<&'static str>() {
-                Some(s) => write!(f, "unexpected panic: {}", s),
-                None => write!(f, "unknown unexpected panic"),
-            },
-        }
-    }
-}
-
-impl From<SignalProtocolError> for SignalFfiError {
-    fn from(e: SignalProtocolError) -> SignalFfiError {
-        SignalFfiError::Signal(e)
-    }
-}
-
-impl From<AesGcmSivError> for SignalFfiError {
-    fn from(e: AesGcmSivError) -> SignalFfiError {
-        SignalFfiError::AesGcmSiv(e)
-    }
-}
-
-pub fn run_ffi_safe<F: FnOnce() -> Result<(), SignalFfiError> + std::panic::UnwindSafe>(
-    f: F,
-) -> *mut SignalFfiError {
-    let result = match std::panic::catch_unwind(f) {
-        Ok(Ok(())) => Ok(()),
-        Ok(Err(e)) => Err(e),
-        Err(r) => Err(SignalFfiError::UnexpectedPanic(r)),
-    };
-
-    match result {
-        Ok(()) => std::ptr::null_mut(),
-        Err(e) => Box::into_raw(Box::new(e)),
     }
 }
