+++ rust/bridge/jni/Cargo.toml
@@ -19,6 +19,7 @@ crate-type = ["cdylib"]
 [dependencies]
 libsignal-protocol-rust = { path = "../../protocol" }
 aes-gcm-siv = { path = "../../aes-gcm-siv" }
+libsignal-bridge = { path = "../shared", features = ["jni"] }
 async-trait = "0.1.41"
 futures = "0.3.7"
 jni = "0.17"
