+++ rust/bridge/jni/src/lib.rs
@@ -13,6 +13,7 @@ use jni::JNIEnv;
 use std::convert::TryFrom;
 
 use aes_gcm_siv::Aes256GcmSiv;
+use libsignal_bridge::*;
 use libsignal_protocol_rust::*;
 
 mod util;
@@ -41,7 +42,7 @@ pub unsafe extern "C" fn Java_org_signal_client_internal_Native_ProtocolAddress_
     })
 }
 
-jni_fn_destroy!(Java_org_signal_client_internal_Native_ProtocolAddress_1Destroy destroys ProtocolAddress);
+// jni_fn_destroy!(Java_org_signal_client_internal_Native_ProtocolAddress_1Destroy destroys ProtocolAddress);
 
 jni_fn_get_jstring!(Java_org_signal_client_internal_Native_ProtocolAddress_1Name(ProtocolAddress) using
                     |p: &ProtocolAddress| Ok(p.name().to_string()));
